src/VumbaSoft.AdventureWorks.Controllers/Production/ProductInventories/ProductInventoriesController.cs
using Microsoft.AspNetCore.Mvc;
using VumbaSoft.AdventureWorks.Objects;
using VumbaSoft.AdventureWorks.Services;
using VumbaSoft.AdventureWorks.Validators;
using System;

namespace VumbaSoft.AdventureWorks.Controllers.Production
{
    [Area("Production")]
    public class ProductInventoriesController : ValidatedController<IProductInventoryValidator, IProductInventoryService>
    {
        public ProductInventoriesController(IProductInventoryValidator validator, IProductInventoryService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create(ProductInventoryView inventory)
        {
            if (!Validator.CanCreate(inventory))
                return View(inventory);

            Service.Create(inventory);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(Int32 id)
        {
            return NotEmptyView(Service.Get<ProductInventoryView>(id));
        }

        [HttpGet]
        public ActionResult Edit(Int32 id)
        {
            return NotEmptyView(Service.Get<ProductInventoryView>(id));
        }

        [HttpPost]
        public ActionResult Edit(ProductInventoryView inventory)
        {
            if (!Validator.CanEdit(inventory))
                return View(inventory);

            Service.Edit(inventory);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(Int32 id)
        {
            return NotEmptyView(Service.Get<ProductInventoryView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(Int32 id)
        {
            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
